requirements.txt
asyncio==3.4.3
autopep8>=1.5.4
boto3==1.15.11
cantools==32.19.0
dnspython>=1.16.0
flake8>=3.8.4
paho-mqtt==1.5.0
pycodestyle>=2.6.0
pyflakes==2.2.0
pylint==1.7.2
pytest>=4.6.11
python-can==3.3.3
python-dotenv==0.14.0
websockets>=5.0
